+++ glossary.js
@@ -120,6 +120,45 @@ if (containerElement) {
         document.querySelector('body').classList.toggle('overflow-hidden', false);
     }
 
+    function waitForElement(elementId) {
+        return new Promise(resolve => {
+            if (document.getElementById(elementId)) {
+                return resolve(document.getElementById(elementId));
+            }
+
+            const observer = new MutationObserver(mutations => {
+                if (document.getElementById(elementId)) {
+                    resolve(document.getElementById(elementId));
+                    observer.disconnect();
+                }
+            });
+
+            observer.observe(document.body, {
+                childList: true,
+                subtree: true
+            });
+        });
+    }
+
+    const untilAsync = async (fn, time = 1000, wait = 10000) => {
+        const startTime = new Date().getTime();
+        for (; ;) {
+            try {
+                if (await fn()) {
+                    return true;
+                }
+            } catch (e) {
+                throw e;
+            }
+
+            if (new Date().getTime() - startTime > wait) {
+                throw new Error('Timed out waiting for condition to become true.');
+            } else {
+                await new Promise((resolve) => setTimeout(resolve, time));
+            }
+        }
+    };
+
     app.ports.allowBackgroundScrolling.subscribe(() => {
         allowBackgroundScrolling();
     });
@@ -137,6 +176,21 @@ if (containerElement) {
             // Do this just in case, as it seems that there might be situations where the background is left "locked".
             allowBackgroundScrolling();
         }
+    });
+
+    app.ports.giveSearchFieldFocusOnceItIsPresent.subscribe((elementId) => {
+        waitForElement(elementId).then(async (element) => {
+            try {
+                await untilAsync(() => {
+                    if (element)
+                        element.focus();
+
+                    document.activeElement === document.getElementById(elementId);
+                }, 50, 2000)
+            } catch (e) {
+                // ignore
+            }
+        });
     });
 
     app.ports.scrollSearchResultIntoView.subscribe((elementId) => {