+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This applies from version 0.5.0 onwards, as some versions before that have broke
 ### Fixed
 
 - Fix layout bug that was causing a space to appear before each comma under "See also".
+- Give search field focus more reliably.
 
 ### Removed